requirements.txt
flask==3.1.0
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
apscheduler==3.10.4
openai==1.58.1
google-generativeai>=0.8.5
lxml==5.3.1
python-dotenv==1.0.1
duckduckgo-search>=6.0.0
apify-client>=1.8.0
